yolo-server/requirements.txt
flask
flask-cors
ultralytics==8.3.29
pillow
numpy
waitress
